Use a mutex when walking the global CWD.
Since the findFirst / findNext operations cause seek operations on the file,
it's unsafe to do multiple find operations on the same directory handle at the
same time. This could happen if multiple fopen calls are done at once. By using
a global mutex, one of the operations is forced to wait until the active one is
complete. The mutex doesn't use true atomic compare and swap operations, so the
docs are also updated to mention that the code is not reentrant.